Move Torrent class to the Entity directory
